# Apply `functools.lru_cache` to `RegexFSM` to Improve `CFGFSM` performance