@lapp0 2024-02-13
@rlouf this and https://github.com/outlines-dev/outlines/pull/622 https://github.com/outlines-dev/outlines/pull/623 are relatively simple changes. They cumulatively increase CFGFSM performance 20-fold. I would like to merge these prior to my introduction of more grammars (https://github.com/outlines-dev/outlines/pull/587).

Please review when you have a chance.